requirements.txt
opencv-python
mediapipe
python-dotenv
numpy
scipy

